fix: fix wait time after minting
