+++ examples/scripts/txTransformTransfer.ts
@@ -2,10 +2,12 @@ import { Commitment, UserFactory } from "../../libs";
 import { createZkpKeys } from "../../libs";
 import { randomL2TokenAddress, randomSalt } from "../../libs";
 import { config } from "./appConfig";
+import getBalance from "./utils/balance";
 import makeBlock from "./utils/blocks";
 
 const main = async () => {
   let userSender;
+  const value = 6;
 
   try {
     // # 1 Create an instance of User
@@ -25,14 +27,14 @@ const main = async () => {
       zkpPublicKey: userSender.zkpKeys.zkpPublicKey.map((x) => BigInt(x)),
       ercAddress: BigInt(tokenContractAddress),
       tokenId: BigInt(1),
-      value: BigInt(6),
+      value: BigInt(value),
       salt: BigInt(salt),
     });
     const inputTokens = [
       {
         id: 1,
         address: tokenContractAddress,
-        value: 6,
+        value,
         salt,
         commitmentHash: `0x${BigInt(commitment.hash._hex)
           .toString(16)
@@ -43,7 +45,7 @@ const main = async () => {
       {
         id: 2,
         address: tokenContractAddress,
-        value: 6,
+        value,
         salt,
       },
     ];
@@ -51,16 +53,22 @@ const main = async () => {
     const { txHashL2: txHashL2mint } = await userSender.mintL2Token({
       tokenContractAddress,
       tokenId: 1,
-      value: "6",
+      value: value.toString(),
       salt,
       feeWei: "0",
     });
 
     console.log(">>>>> Mint transaction hash L2", txHashL2mint);
-    await makeBlock(10000);
+    await makeBlock();
 
-    console.log(">>>>> Wait some time for minted token to be in a block");
-    await new Promise((resolve) => setTimeout(resolve, 20000));
+    console.log(">>>>> Wait for minted token to be in a block");
+
+    let balanceToken = await getBalance(userSender, tokenContractAddress);
+    while (balanceToken !== value) {
+      await new Promise((resolve) => setTimeout(resolve, 1000));
+      balanceToken = await getBalance(userSender, tokenContractAddress);
+    }
+
     // # 3 Make transformTransfer
     // For this example, we generate a L2 address to receive the transformTransfer
     const { zkpKeys } = await createZkpKeys(config.clientApiUrl);
